requirements/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
duckdb>=0.9.0
structlog>=23.0.0
python-telegram-bot>=20.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
